#28 Sound plays, but at wrong speed
The sound plays when a button is clicked, but the sample rate is screwed
up so that it sounds really bad. I recall us having problems reading in
existing WAV files into arrays way back when Isabel was working on this
code, so this is a major bit of code in a utility class that is simply
not available at the moment, and has to be added first.

Perhaps more seriously, the sound array inputs do not seem to be
meaningfully influencing the animation. When I disable all but the sound
amplitude input to the CPPN, there is (almost?) no animation at all. Is
this because I am using such a short clip of the sound file now, or is
there a larger issue?